pyproject.toml
[project]
name = "justdoit"
version = "0.1.0"
description = "Zero-dependency Python 3 ASCII art CLI with block, slim, and figlet fonts"
authors = [{ name = "Jonny Galloway", email = "NumberOne@co3dex.com" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

# Core has no required dependencies — pure stdlib.
# Optional extras gate on Pillow for TTF/OTF support.
dependencies = []

[project.optional-dependencies]
ttf = ["Pillow>=10.0"]
dev = [
    "pytest>=8.0",
    "Pillow>=10.0",
]

[project.scripts]
justdoit = "justdoit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["justdoit"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["E221"]  # allow aligned assignments (used throughout for dict/list readability)

[dependency-groups]
dev = [
    "pytest>=8.0",
    "Pillow>=10.0",
    "numpy>=1.24",
]
